Data.Total.MonoidMap

Description

Total monoidal map type with support for semigroup and monoid subclasses.

Type

data MonoidMap k v

Models a total relation from unique keys to monoidal values.

The mapping from keys to values is total: every possible key of type k is associated with a corresponding value of type v:

get :: (Ord k, Monoid v) => k -> MonoidMap k v -> v

By default, every key in an empty map corresponds to a value of mempty:

∀ k. get k empty == mempty

Comparison with partial map types

The MonoidMap type differs from the standard Map type in how it relates keys to values:

Map k v:

relates keys of type k to values of type Maybe v.

MonoidMap k v:

relates keys of type k to values of type v.

This becomes evident if we compare the type signatures of operations to query a key for its value, for both types:

      Map.lookup ::             k ->       Map k v -> Maybe v
MonoidMap.get    :: Monoid v => k -> MonoidMap k v ->       v

For unconstrained value types, using Maybe makes it possible to signal the presence or absence of a value for a particular key.

However, monoidal types have a natural way to represent empty values: the mempty constant, which represents the identity element of a monoid.

Consequently, using a standard Map with a monoidal value type gives rise to two distinct representations for missing or empty values:

Map.lookup k m

Nothing 	Map m has no entry for key k.

Just mempty 	Map m does have an entry for key k, but the value is empty.

In constrast, the MonoidMap type provides a single, canonical representation for empty values, according to the following mapping:

Map.lookup k m 	 	 MonoidMap.get k m

Nothing 	 ⟼ 	 mempty

Just v 	 v == mempty 	 ⟼ 	 mempty

Just v 	 v /= mempty 	 ⟼ 	 v

Internal data structure

Internally, the MonoidMap type uses a sparse Map data structure to store its key-value mappings, and only stores values that are not equal to mempty.

Values that are equal to mempty are automatically garbage collected, and never included in the internal data structure.

As a result, the internal data structure is always in a canonical form.

Instances of Semigroup and Monoid

This module provides a Semigroup instance that uses the (<>) operator to combines values for matching keys, satisfying the following property:

get k (m1 <> m2) == get k m1 <> get k m2

The Monoid instance satisfies the following property for all keys:

get k mempty == mempty

Subclasses of Semigroup and Monoid

This module also provides instances for several subclasses of Semigroup and Monoid.

In general, these instances are defined analogously to the Semigroup instance, where binary operations on pairs of maps are defined in terms of their application to pairs of values for matching keys.

For example, if subclass C defines a binary operation f of the form:

class Semigroup a => C a where
   f :: a -> a -> a

Then the result of applying f to maps m1 and m2 will typically satisfy a property of the following form:

∀ k m1 m2. get k (f m1 m2) == f (get k m1) (get k m2)

Examples

Expand

The commonPrefix function from LeftGCDMonoid:

get k (commonPrefix m1 m2)
    == commonPrefix (get k m1) (get k m2)

The commonSuffix function from RightGCDMonoid:

get k (commonSuffix m1 m2)
    == commonSuffix (get k m1) (get k m2)

The overlap function from OverlappingGCDMonoid:

get k (overlap m1 m2)
    == overlap (get k m1) (get k m2)

Constraints on value types

MonoidMap operations generally require the value type to be an instance of MonoidNull, which provides a convenient way to test whether or not a value is mempty.

Types that are instances of both MonoidNull and Eq satisfy the following equivalence:

null v == (v == mempty)

However, MonoidMap operations generally do not require that value types are instances of Eq.

Justification

The set of monoidal types that admit a MonoidNull instance is strictly larger than the set of monoidal types that admit an Eq instance.

For any type v that is an instance of both Eq and Monoid, it is always possible to define a MonoidNull instance:

instance MonoidNull v where
    null = (== mempty)

However, there are monoidal types for which it is possible to define a MonoidNull instance, but not practical (or possible) to define a lawful Eq instance.

For example, consider the following type:

Maybe (String -> Sum Natural)

Requiring a MonoidNull constraint instead of an Eq constraint allows MonoidMap to be usable with a greater range of monoidal value types.

Construction

empty :: MonoidMap k v

The empty MonoidMap.

Satisfies the following property:

get k empty == mempty

Provides the definition of mempty for the MonoidMap instance of Monoid.

fromList :: (Ord k, MonoidNull v) => [(k, v)] -> MonoidMap k v

Constructs a MonoidMap from a list of key-value pairs.

If the list contains more than one value for the same key, values are combined together with (<>) in the order that they appear in the list.

Satisfies the following property for all possible keys k:

get k (fromList kvs) == foldMap snd (filter ((== k) . fst) kvs)

Satisfies the following round-trip property:

fromList (toList m) == m

fromListWith

Arguments

:: (Ord k, MonoidNull v)

=> (v -> v -> v)
Combination function with which to combine values for duplicate keys.

-> [(k, v)]

-> MonoidMap k v

Constructs a MonoidMap from a list of key-value pairs.

If the list contains more than one value for the same key, values are combined together with the given combination function in the order that they appear in the list.

fromMap :: MonoidNull v => Map k v -> MonoidMap k v

Constructs a MonoidMap from an ordinary Map.

Satisfies the following property:

get k (fromMap m) == findWithDefault mempty k m

singleton :: (Ord k, MonoidNull v) => k -> v -> MonoidMap k v

Constructs a MonoidMap from a single key-value pair.

Satisfies the following property:

get k (singleton k v) == v

Nullifying the value for key k produces an empty map:

nullify k (singleton k v) == empty

Deconstruction

toList :: MonoidMap k v -> [(k, v)]

Converts a MonoidMap to a list of key-value pairs.

The result only includes entries with values that are not null.

Satisfies the following property:

fromList (toList m) == m

toMap :: MonoidMap k v -> Map k v

Converts a MonoidMap to an ordinary Map.

The result only includes entries with values that are not null.

Satisfies the following property:

fromMap (toMap m) == m

Lookup

get :: (Ord k, Monoid v) => k -> MonoidMap k v -> v

Gets the value associated with the given key.

Satisfies the following property:

get k (set k v m) == v

Modification

set :: (Ord k, MonoidNull v) => k -> v -> MonoidMap k v -> MonoidMap k v

Sets the value associated with the given key.

Satisfies the following property:

get k (set k v m) == v

adjust :: (Ord k, MonoidNull v) => (v -> v) -> k -> MonoidMap k v -> MonoidMap k v

Adjusts the value associated with the given key.

Satisfies the following property:

adjust f k m == set k (f (get k m)) m

nullify :: Ord k => k -> MonoidMap k v -> MonoidMap k v

Sets the value associated with the given key to mempty.

Satisfies the following property:

get k (nullify k m) == mempty

Membership

null :: MonoidMap k v -> Bool

Returns True if (and only if) all values in the map are null.

Satisfies the following property:

null m == (∀ k. nullKey k m)

Provides the definition of null for the MonoidMap instance of MonoidNull.

nullKey :: Ord k => k -> MonoidMap k v -> Bool

Returns True if (and only if) the given key is associated with a value that is null.

Satisfies the following property:

nullKey k m == null (get k m)

nonNull :: MonoidMap k v -> Bool

Returns True if (and only if) the map contains at least one value that is not null.

Satisfies the following property:

nonNull m == (∃ k. nonNullKey k m)

nonNullCount :: MonoidMap k v -> Int

Returns a count of all values in the map that are not null.

Satisfies the following property:

nonNullCount m == size (nonNullKeys m)

nonNullKey :: Ord k => k -> MonoidMap k v -> Bool

Returns True if (and only if) the given key is associated with a value that is not null.

Satisfies the following property:

nonNullKey k m == not (null (get k m))

nonNullKeys :: MonoidMap k v -> Set k

Returns the set of keys associated with values that are not null.

Satisfies the following property:

k `member` (nonNullKeys m) == nonNullKey k m

Slicing

take :: Int -> MonoidMap k v -> MonoidMap k v

Takes a slice from a map.

This function takes a given number of non-null entries from a map, producing a new map from the entries that were taken.

Entries are taken in key order, beginning with the smallest keys.

Satifies the following property:

take n == fromList . take n . toList

drop :: Int -> MonoidMap k v -> MonoidMap k v

Drops a slice from a map.

This function drops a given number of non-null entries from a map, producing a new map from the entries that remain.

Entries are dropped in key order, beginning with the smallest keys.

Satifies the following property:

drop n == fromList . drop n . toList

splitAt :: Int -> MonoidMap k a -> (MonoidMap k a, MonoidMap k a)

Splits a map into two slices.

This function is equivalent to a combination of take and drop:

splitAt n m == (take n m, drop n m)

The resulting maps can be combined to reproduce the original map:

splitAt n m &
    \(m1, m2) -> m1 <> m2 == m

The resulting maps have disjoint sets of non-null entries:

splitAt n m &
    \(m1, m2) -> disjoint (nonNullKeys m1) (nonNullKeys m2)

Filtering

filter :: (k -> v -> Bool) -> MonoidMap k v -> MonoidMap k v

Filters the non-null entries of a map according to a predicate on keys and values.

The result includes just the subset of non-null entries that satisfy the predicate.

Satisfies the following property:

toList (filter f m) == filter (uncurry f) (toList m)

filterKeys :: (k -> Bool) -> MonoidMap k v -> MonoidMap k v

Filters the non-null entries of a map according to a predicate on keys.

The result includes just the subset of non-null entries that satisfy the predicate.

Satisfies the following property:

filterKeys f m == filter (\k _ -> f k) m

filterValues :: (v -> Bool) -> MonoidMap k v -> MonoidMap k v

Filters the non-null entries of a map according to a predicate on values.

The result includes just the subset of non-null entries that satisfy the predicate.

Satisfies the following property:

filterValues f m == filter (\_ v -> f v) m

Partitioning

partition :: (k -> v -> Bool) -> MonoidMap k v -> (MonoidMap k v, MonoidMap k v)

Partitions the non-null entries of a map according to a predicate on keys and values.

The first map includes the subset of non-null entries that satisfy the predicate, and the second map includes the subset of non-null entries that fail the predicate.

Satisfies the following property:

partition f m == (filter f m, filter ((fmap . fmap) not f) m)

The resulting maps can be combined to reproduce the original map:

partition f m &
    \(m1, m2) -> m1 <> m2 == m

The resulting maps have disjoint sets of non-null entries:

partition f m &
    \(m1, m2) -> disjoint (nonNullKeys m1) (nonNullKeys m2)

partitionKeys :: (k -> Bool) -> MonoidMap k v -> (MonoidMap k v, MonoidMap k v)

Partitions the non-null entries of a map according to a predicate on keys.

The first map includes the subset of non-null entries that satisfy the predicate, and the second map includes the subset of non-null entries that fail the predicate.

Satisfies the following property:

partitionKeys f m == (filterKeys f m, filterKeys (not . f) m)

The resulting maps can be combined to reproduce the original map:

partitionKeys f m &
    \(m1, m2) -> m1 <> m2 == m

The resulting maps have disjoint sets of non-null entries:

partitionKeys f m &
    \(m1, m2) -> disjoint (nonNullKeys m1) (nonNullKeys m2)

partitionValues :: (v -> Bool) -> MonoidMap k v -> (MonoidMap k v, MonoidMap k v)

Partitions the non-null entries of a map according to a predicate on values.

The first map includes the subset of non-null entries that satisfy the predicate, and the second map includes the subset of non-null entries that fail the predicate.

Satisfies the following property:

partitionValues f m == (filterValues f m, filterValues (not . f) m)

The resulting maps can be combined to reproduce the original map:

partitionValues f m &
    \(m1, m2) -> m1 <> m2 == m

The resulting maps have disjoint sets of non-null entries:

partitionValues f m &
    \(m1, m2) -> disjoint (nonNullKeys m1) (nonNullKeys m2)

Mapping

map :: (Ord k2, MonoidNull v2) => (k1 -> k2) -> (v1 -> v2) -> MonoidMap k1 v1 -> MonoidMap k2 v2

Maps over the keys and values of a MonoidMap.

Satisfies the following property:

map f g == fromList . fmap (bimap f g) . toList

mapWith :: (Ord k2, MonoidNull v2) => (v2 -> v2 -> v2) -> (k1 -> k2) -> (v1 -> v2) -> MonoidMap k1 v1 -> MonoidMap k2 v2

Maps over the keys and values of a MonoidMap.

Satisfies the following property:

mapWith c f g == fromListWith c . fmap (bimap f g) . toList

mapKeys :: (Ord k2, MonoidNull v) => (k1 -> k2) -> MonoidMap k1 v -> MonoidMap k2 v

Maps over the keys of a MonoidMap.

Satisfies the following property:

mapKeys f == fromList . fmap (first f) . toList

mapKeysWith :: (Ord k2, MonoidNull v) => (v -> v -> v) -> (k1 -> k2) -> MonoidMap k1 v -> MonoidMap k2 v

Maps over the keys of a MonoidMap.

Satisfies the following property:

mapKeysWith c f == fromListWith c . fmap (first f) . toList

mapValues :: MonoidNull v2 => (v1 -> v2) -> MonoidMap k v1 -> MonoidMap k v2

Applies a function to all non-null values of a MonoidMap.

Satisfies the following properties for all functions f:

(get k m == mempty) ==> (get k (mapValues f m) == mempty     )
(get k m /= mempty) ==> (get k (mapValues f m) == f (get k m))

If function f is a monoid homomorphism, then the mapping is total for all possible keys k:

(f mempty == mempty) ==>
    (∀ k. get k (mapValues f m) == f (get k m)))

Association

append :: (Ord k, MonoidNull v) => MonoidMap k v -> MonoidMap k v -> MonoidMap k v

Appends a pair of maps together.

Uses the Semigroup operator (<>) to append each value in the first map to its matching value in the second map.

Satisfies the following property for all keys k:

get k (append m1 m2) == get k m1 <> get k m2

This function provides the definition of (<>) for the MonoidMap instance of Semigroup.

Examples

Expand

With String values:

>>> m1 = fromList [(1, "abc"), (2, "ij" ), (3, "p"  )            ]
>>> m2 = fromList [            (2, "  k"), (3,  "qr"), (4, "xyz")]
>>> m3 = fromList [(1, "abc"), (2, "ijk"), (3, "pqr"), (4, "xyz")]

>>> append m1 m2 == m3
True

With Sum Natural values:

>>> m1 = fromList [("a", 4), ("b", 2), ("c", 1)          ]
>>> m2 = fromList [          ("b", 1), ("c", 2), ("d", 4)]
>>> m3 = fromList [("a", 4), ("b", 3), ("c", 3), ("d", 4)]

>>> append m1 m2 == m3
True

Subtraction

minus :: (Ord k, MonoidNull v, Group v) => MonoidMap k v -> MonoidMap k v -> MonoidMap k v

Performs group subtraction of the second map from the first.

Uses the Group subtraction operator (~~) to subtract each value in the second map from its matching value in the first map.

Satisfies the following property:

get k (m1 `minus` m2) == get k m1 ~~ get k m2

This function provides the definition of (~~) for the MonoidMap instance of Group.

Examples

Expand

With Sum Integer values, this function performs normal integer subtraction of matching values:

>>> m1 = fromList [("a", (-1)), ("b",   0 ), ("c", 1)]
>>> m2 = fromList [("a",   1 ), ("b",   1 ), ("c", 1)]
>>> m3 = fromList [("a", (-2)), ("b", (-1)), ("c", 0)]

>>> m1 `minus` m2 == m3
True

>>> m1 = fromList [("a", (-1)), ("b",   0 ), ("c",   1 )]
>>> m2 = fromList [("a", (-1)), ("b", (-1)), ("c", (-1))]
>>> m3 = fromList [("a",   0 ), ("b",   1 ), ("c",   2 )]

>>> m1 `minus` m2 == m3
True

minusMaybe :: (Ord k, MonoidNull v, Reductive v) => MonoidMap k v -> MonoidMap k v -> Maybe (MonoidMap k v)

Performs reductive subtraction of the second map from the first.

Uses the Reductive subtraction operator (</>) to subtract each value in the second map from its matching value in the first map.

This function produces a result if (and only if) for all possible keys k, it is possible to subtract the value for k in the second map from the value for k in the first map:

isJust (m1 `minusMaybe` m2)
    == (∀ k. isJust (get k m1 </> get k m2))

Otherwise, this function returns Nothing.

This function satisfies the following property:

all
   (\r -> Just (get k r) == get k m1 </> get k m2)
   (m1 `minusMaybe` m2)

This function provides the definition of (</>) for the MonoidMap instance of Reductive.

Examples

Expand

With Set Natural values, this function performs set subtraction of matching values, succeeding if (and only if) each value from the second map is a subset of its matching value from the first map:

f xs = fromList (fromList <$> xs)

>>> m1 = f [("a", [0,1,2]), ("b", [0,1,2])]
>>> m2 = f [("a", [     ]), ("b", [0,1,2])]
>>> m3 = f [("a", [0,1,2]), ("b", [     ])]

>>> m1 `minusMaybe` m2 == Just m3
True

>>> m1 = f [("a", [0,1,2]), ("b", [0,1,2]), ("c", [0,1,2])]
>>> m2 = f [("a", [0    ]), ("b", [  1  ]), ("c", [    2])]
>>> m3 = f [("a", [  1,2]), ("b", [0,  2]), ("c", [0,1  ])]

>>> m1 `minusMaybe` m2 == Just m3
True

>>> m1 = f [("a", [0,1,2    ]), ("b", [0,1,2    ]), ("c", [0,1,2    ])]
>>> m2 = f [("a", [    2,3,4]), ("b", [  1,2,3,4]), ("c", [0,1,2,3,4])]

>>> m1 `minusMaybe` m2 == Nothing
True

With Sum Natural values, this function performs ordinary subtraction of matching values, succeeding if (and only if) each value from the second map is less than or equal to its matching value from the first map:

>>> m1 = fromList [("a", 2), ("b", 3), ("c", 5), ("d", 8)]
>>> m2 = fromList [("a", 0), ("b", 0), ("c", 0), ("d", 0)]
>>> m3 = fromList [("a", 2), ("b", 3), ("c", 5), ("d", 8)]

>>> m1 `minusMaybe` m2 == Just m3
True

>>> m1 = fromList [("a", 2), ("b", 3), ("c", 5), ("d", 8)]
>>> m2 = fromList [("a", 1), ("b", 2), ("c", 3), ("d", 5)]
>>> m3 = fromList [("a", 1), ("b", 1), ("c", 2), ("d", 3)]

>>> m1 `minusMaybe` m2 == Just m3
True

>>> m1 = fromList [("a", 2), ("b", 3), ("c", 5), ("d", 8)]
>>> m2 = fromList [("a", 2), ("b", 3), ("c", 5), ("d", 8)]
>>> m3 = fromList [("a", 0), ("b", 0), ("c", 0), ("d", 0)]

>>> m1 `minusMaybe` m2 == Just m3
True

>>> m1 = fromList [("a", 2), ("b", 3), ("c", 5), ("d", 8)]
>>> m2 = fromList [("a", 3), ("b", 3), ("c", 5), ("d", 8)]

>>> m1 `minusMaybe` m2 == Nothing
True

monus :: (Ord k, MonoidNull v, Monus v) => MonoidMap k v -> MonoidMap k v -> MonoidMap k v

Performs monus subtraction of the second map from the first.

Uses the Monus subtraction operator (<\>) to subtract each value in the second map from its matching value in the first map.

Satisfies the following property:

get k (m1 `monus` m2) == get k m1 <\> get k m2

This function provides the definition of (<\>) for the MonoidMap instance of Monus.

Examples

Expand

With Set Natural values, this function performs set subtraction of matching values:

f xs = fromList (fromList <$> xs)

>>> m1 = f [("a", [0,1,2]), ("b", [0,1,2])]
>>> m2 = f [("a", [     ]), ("b", [0,1,2])]
>>> m3 = f [("a", [0,1,2]), ("b", [     ])]

>>> m1 `monus` m2 == m3
True

>>> m1 = f [("a", [0,1,2]), ("b", [0,1,2]), ("c", [0,1,2])]
>>> m2 = f [("a", [0    ]), ("b", [  1  ]), ("c", [    2])]
>>> m3 = f [("a", [  1,2]), ("b", [0,  2]), ("c", [0,1  ])]

>>> m1 `monus` m2 == m3
True

>>> m1 = f [("a", [0,1,2    ]), ("b", [0,1,2    ]), ("c", [0,1,2    ])]
>>> m2 = f [("a", [    2,3,4]), ("b", [  1,2,3,4]), ("c", [0,1,2,3,4])]
>>> m3 = f [("a", [0,1      ]), ("b", [0        ]), ("c", [         ])]

>>> m1 `monus` m2 == m3
True

With Sum Natural values, this function performs truncated subtraction of matching values:

>>> m1 = fromList [("a", 0), ("b", 1), ("c", 2), ("d", 3)]
>>> m2 = fromList [("a", 0), ("b", 0), ("c", 0), ("d", 0)]
>>> m3 = fromList [("a", 0), ("b", 1), ("c", 2), ("d", 3)]

>>> m1 `monus` m2 == m3
True

>>> m1 = fromList [("a", 0), ("b", 1), ("c", 2), ("d", 3)]
>>> m2 = fromList [("a", 1), ("b", 1), ("c", 1), ("d", 1)]
>>> m3 = fromList [("a", 0), ("b", 0), ("c", 1), ("d", 2)]

>>> m1 `monus` m2 == m3
True

>>> m1 = fromList [("a", 0), ("b", 1), ("c", 2), ("d", 3)]
>>> m2 = fromList [("a", 2), ("b", 2), ("c", 2), ("d", 2)]
>>> m3 = fromList [("a", 0), ("b", 0), ("c", 0), ("d", 1)]

>>> m1 `monus` m2 == m3
True

>>> m1 = fromList [("a", 0), ("b", 1), ("c", 2), ("d", 3)]
>>> m2 = fromList [("a", 4), ("b", 4), ("c", 4), ("d", 4)]
>>> m3 = fromList [("a", 0), ("b", 0), ("c", 0), ("d", 0)]

>>> m1 `monus` m2 == m3
True

Inversion

invert :: (Ord k, MonoidNull v, Group v) => MonoidMap k v -> MonoidMap k v

Inverts every value in a map.

Applies the Group method invert to every value in a map.

Satisfies the following property:

get k (invert m) == invert (get k m)

This function provides the definition of invert for the MonoidMap instance of Group.

Examples

Expand

With Sum Integer values, this function performs negation of values:

>>> m1 = fromList [("a", (-1)), ("b", 0), ("c",   1) ]
>>> m2 = fromList [("a",   1 ), ("b", 0), ("c", (-1))]

>>> negate m1 == m2
True

Exponentiation

power :: (Integral i, MonoidNull v, Group v) => MonoidMap k v -> i -> MonoidMap k v

Performs exponentiation of every value in a map.

Uses the Group exponentiation method pow to raise every value in a map to the power of the given exponent.

Satisfies the following property:

get k (m `power` i) == get k m `pow` i

This function provides the definition of pow for the MonoidMap instance of Group.

Examples

Expand

With Sum Natural values, this function performs ordinary multiplication of all values by the given exponent:

>>> m1 = fromList [("a", 0), ("b", 1), ("c", 2), ("d", 3)]
>>> m2 = fromList [("a", 0), ("b", 2), ("c", 4), ("d", 6)]

>>> m1 `power` 2 == m2
True

>>> m1 = fromList [("a", 0), ("b",   1 ), ("c",   2 ), ("d",   3 )]
>>> m2 = fromList [("a", 0), ("b", (-1)), ("c", (-2)), ("d", (-3))]

>>> m1 `power` (-1) == m2
True

Prefixes

isPrefixOf :: (Ord k, Monoid v, LeftReductive v) => MonoidMap k v -> MonoidMap k v -> Bool

Indicates whether or not the first map is a prefix of the second.

MonoidMap m1 is a prefix of MonoidMap m2 if (and only if) for all possible keys k, the value for k in m1 is a prefix of the value for k in m2:

m1 `isPrefixOf` m2 == (∀ k. get k m1 `isPrefixOf` get k m2)

This function provides the definition of isPrefixOf for the MonoidMap instance of LeftReductive.

Examples

Expand

With String values:

>>> m1 = fromList [(1, "a"  ), (2, "p"  ), (3, "x"  )]
>>> m2 = fromList [(1, "abc"), (2, "pqr"), (3, "xyz")]
>>> m1 `isPrefixOf` m2
True

>>> m1 = fromList [            (2, "p"  )            ]
>>> m2 = fromList [(1, "abc"), (2, "pqr"), (3, "xyz")]
>>> m1 `isPrefixOf` m2
True

>>> m1 = fromList [(1, "abc"), (2, "p"  ), (3, "x"  )]
>>> m2 = fromList [(1, "a"  ), (2, "pqr"), (3, "xyz")]
>>> m1 `isPrefixOf` m2
False

With Sum Natural values:

>>> m1 = fromList [("a", 1), ("b", 1), ("c", 1)]
>>> m2 = fromList [("a", 2), ("b", 4), ("c", 8)]
>>> m1 `isPrefixOf` m2
True

>>> m1 = fromList [          ("b", 1)          ]
>>> m2 = fromList [("a", 2), ("b", 4), ("c", 8)]
>>> m1 `isPrefixOf` m2
True

>>> m1 = fromList [("a", 2), ("b", 1), ("c", 1)]
>>> m2 = fromList [("a", 1), ("b", 4), ("c", 8)]
>>> m1 `isPrefixOf` m2
False

stripPrefix :: (Ord k, MonoidNull v, LeftReductive v) => MonoidMap k v -> MonoidMap k v -> Maybe (MonoidMap k v)

Strips a prefix from a MonoidMap.

If map m1 is a prefix of map m2, then stripPrefix m1 m2 will produce a reduced map where prefix m1 is stripped from m2.

Properties

The stripPrefix function, when applied to maps m1 and m2, produces a result if (and only if) m1 is a prefix of m2:

isJust (stripPrefix m1 m2) == m1 `isPrefixOf` m2

The value for any key k in the result is identical to the result of stripping the value for k in map m1 from the value for k in map m2:

all
   (\r -> Just (get k r) == stripPrefix (get k m1) (get k m2))
   (stripPrefix m1 m2)

If we append prefix m1 to the left-hand side of the result, we can always recover the original map m2:

all
   (\r -> m1 <> r == m2)
   (stripPrefix m1 m2)

This function provides the definition of stripPrefix for the MonoidMap instance of LeftReductive.

Examples

Expand

With String values:

>>> m1 = fromList [(1, ""   ), (2, "i"  ), (3, "pq" ), (4, "xyz")]
>>> m2 = fromList [(1, "abc"), (2, "ijk"), (3, "pqr"), (4, "xyz")]
>>> m3 = fromList [(1, "abc"), (2,  "jk"), (3,   "r"), (4,    "")]

>>> stripPrefix m1 m2 == Just m3
True

>>> stripPrefix m2 m1 == Nothing
True

With Sum Natural values:

>>> m1 = fromList [("a", 0), ("b", 1), ("c", 2), ("d", 3)]
>>> m2 = fromList [("a", 3), ("b", 3), ("c", 3), ("d", 3)]
>>> m3 = fromList [("a", 3), ("b", 2), ("c", 1), ("d", 0)]

>>> stripPrefix m1 m2 == Just m3
True

>>> stripPrefix m2 m1 == Nothing
True

commonPrefix :: (Ord k, MonoidNull v, LeftGCDMonoid v) => MonoidMap k v -> MonoidMap k v -> MonoidMap k v

Finds the greatest common prefix of two maps.

Satisfies the following property:

get k (commonPrefix m1 m2)
    == commonPrefix (get k m1) (get k m2)

This function provides the definition of commonPrefix for the MonoidMap instance of LeftGCDMonoid.

Examples

Expand

With String values:

>>> m1 = fromList [(1, "+++"), (2, "b++"), (3, "cc+"), (4, "ddd")]
>>> m2 = fromList [(1, "---"), (2, "b--"), (3, "cc-"), (4, "ddd")]
>>> m3 = fromList [(1, ""   ), (2, "b"  ), (3, "cc" ), (4, "ddd")]

>>> commonPrefix m1 m2 == m3
True

With Sum Natural values:

>>> m1 = fromList [("a", 0), ("b", 1), ("c", 2), ("d", 3)]
>>> m2 = fromList [("a", 2), ("b", 2), ("c", 2), ("d", 2)]
>>> m3 = fromList [("a", 0), ("b", 1), ("c", 2), ("d", 2)]

>>> commonPrefix m1 m2 == m3
True

stripCommonPrefix :: (Ord k, MonoidNull v, LeftGCDMonoid v) => MonoidMap k v -> MonoidMap k v -> (MonoidMap k v, MonoidMap k v, MonoidMap k v)

Strips the greatest common prefix from a pair of maps.

Given two maps m1 and m2, stripCommonPrefix produces a tuple (p, r1, r2), where:

p is the greatest common prefix of m1 and m2

r1 is the remainder of stripping prefix p from m1

r2 is the remainder of stripping prefix p from m2

The resulting prefix p can be appended to the left-hand side of either remainder r1 or r2 to reproduce either of the original maps m1 or m2 respectively:

stripCommonPrefix m1 m2
   & \(p, r1, _) -> p <> r1 == m1
stripCommonPrefix m1 m2
   & \(p, _, r2) -> p <> r2 == m2

Prefix p is identical to the result of applying commonPrefix to m1 and m2:

stripCommonPrefix m1 m2
   & \(p, _, _) -> p == commonPrefix m1 m2

Remainders r1 and r2 are identical to the results of applying stripPrefix to p and m1 or to p and m2 respectively:

stripCommonPrefix m1 m2
   & \(p, r1, _) -> Just r1 == stripPrefix p m1
stripCommonPrefix m1 m2
   & \(p, _, r2) -> Just r2 == stripPrefix p m2

This function provides the definition of stripCommonPrefix for the MonoidMap instance of LeftGCDMonoid.

Examples

Expand

With String values:

>>> m1 = fromList [(1, "+++"), (2, "a++"), (3, "aa+"), (4, "aaa")]
>>> m2 = fromList [(1, "---"), (2, "a--"), (3, "aa-"), (4, "aaa")]

>>> p  = fromList [(1, ""   ), (2, "a"  ), (3, "aa" ), (4, "aaa")]
>>> r1 = fromList [(1, "+++"), (2,  "++"), (3,   "+"), (4,    "")]
>>> r2 = fromList [(1, "---"), (2,  "--"), (3,   "-"), (4,    "")]

>>> stripCommonPrefix m1 m2 == (p, r1, r2)
True

With Sum Natural values:

>>> m1 = fromList [("a", 0), ("b", 1), ("c", 2), ("d", 3), ("e", 4)]
>>> m2 = fromList [("a", 4), ("b", 3), ("c", 2), ("d", 1), ("e", 0)]

>>> p  = fromList [("a", 0), ("b", 1), ("c", 2), ("d", 1), ("e", 0)]
>>> r1 = fromList [("a", 0), ("b", 0), ("c", 0), ("d", 2), ("e", 4)]
>>> r2 = fromList [("a", 4), ("b", 2), ("c", 0), ("d", 0), ("e", 0)]

>>> stripCommonPrefix m1 m2 == (p, r1, r2)
True

Suffixes

isSuffixOf :: (Ord k, Monoid v, RightReductive v) => MonoidMap k v -> MonoidMap k v -> Bool

Indicates whether or not the first map is a suffix of the second.

MonoidMap m1 is a suffix of MonoidMap m2 if (and only if) for all possible keys k, the value for k in m1 is a suffix of the value for k in m2:

m1 `isSuffixOf` m2 == (∀ k. get k m1 `isSuffixOf` get k m2)

This function provides the definition of isSuffixOf for the MonoidMap instance of RightReductive.

Examples

Expand

With String values:

>>> m1 = fromList [(1,   "c"), (2,   "r"), (3,   "z")]
>>> m2 = fromList [(1, "abc"), (2, "pqr"), (3, "xyz")]
>>> m1 `isSuffixOf` m2
True

>>> m1 = fromList [            (2,   "r")            ]
>>> m2 = fromList [(1, "abc"), (2, "pqr"), (3, "xyz")]
>>> m1 `isSuffixOf` m2
True

>>> m1 = fromList [(1, "abc"), (2,   "r"), (3,   "z")]
>>> m2 = fromList [(1,   "c"), (2, "pqr"), (3, "xyz")]
>>> m1 `isSuffixOf` m2
False

With Sum Natural values:

>>> m1 = fromList [("a", 1), ("b", 1), ("c", 1)]
>>> m2 = fromList [("a", 2), ("b", 4), ("c", 8)]
>>> m1 `isSuffixOf` m2
True

>>> m1 = fromList [          ("b", 1)          ]
>>> m2 = fromList [("a", 2), ("b", 4), ("c", 8)]
>>> m1 `isSuffixOf` m2
True

>>> m1 = fromList [("a", 2), ("b", 1), ("c", 1)]
>>> m2 = fromList [("a", 1), ("b", 4), ("c", 8)]
>>> m1 `isSuffixOf` m2
False

stripSuffix :: (Ord k, MonoidNull v, RightReductive v) => MonoidMap k v -> MonoidMap k v -> Maybe (MonoidMap k v)

Strips a suffix from a MonoidMap.

If map m1 is a suffix of map m2, then stripSuffix m1 m2 will produce a reduced map where suffix m1 is stripped from m2.

Properties

The stripSuffix function, when applied to maps m1 and m2, produces a result if (and only if) m1 is a suffix of m2:

isJust (stripSuffix m1 m2) == m1 `isSuffixOf` m2

The value for any key k in the result is identical to the result of stripping the value for k in map m1 from the value for k in map m2:

all
   (\r -> Just (get k r) == stripSuffix (get k m1) (get k m2))
   (stripSuffix m1 m2)

If we append suffix m1 to the right-hand side of the result, we can always recover the original map m2:

all
   (\r -> r <> m1 == m2)
   (stripSuffix m1 m2)

This function provides the definition of stripSuffix for the MonoidMap instance of RightReductive.

Examples

Expand

With String values:

>>> m1 = fromList [(1,    ""), (2,   "k"), (3,  "qr"), (4, "xyz")]
>>> m2 = fromList [(1, "abc"), (2, "ijk"), (3, "pqr"), (4, "xyz")]
>>> m3 = fromList [(1, "abc"), (2, "ij" ), (3, "p"  ), (4, ""   )]

>>> stripSuffix m1 m2 == Just m3
True

>>> stripSuffix m2 m1 == Nothing
True

With Sum Natural values:

>>> m1 = fromList [("a", 0), ("b", 1), ("c", 2), ("d", 3)]
>>> m2 = fromList [("a", 3), ("b", 3), ("c", 3), ("d", 3)]
>>> m3 = fromList [("a", 3), ("b", 2), ("c", 1), ("d", 0)]

>>> stripSuffix m1 m2 == Just m3
True

>>> stripSuffix m2 m1 == Nothing
True

commonSuffix :: (Ord k, MonoidNull v, RightGCDMonoid v) => MonoidMap k v -> MonoidMap k v -> MonoidMap k v

Finds the greatest common suffix of two maps.

Satisfies the following property:

get k (commonSuffix m1 m2)
    == commonSuffix (get k m1) (get k m2)

This function provides the definition of commonSuffix for the MonoidMap instance of RightGCDMonoid.

Examples

Expand

With String values:

>>> m1 = fromList [(1, "+++"), (2, "++b"), (3, "+cc"), (4, "ddd")]
>>> m2 = fromList [(1, "---"), (2, "--b"), (3, "-cc"), (4, "ddd")]
>>> m3 = fromList [(1,    ""), (2,   "b"), (3,  "cc"), (4, "ddd")]

>>> commonSuffix m1 m2 == m3
True

With Sum Natural values:

>>> m1 = fromList [("a", 0), ("b", 1), ("c", 2), ("d", 3)]
>>> m2 = fromList [("a", 2), ("b", 2), ("c", 2), ("d", 2)]
>>> m3 = fromList [("a", 0), ("b", 1), ("c", 2), ("d", 2)]

>>> commonSuffix m1 m2 == m3
True

stripCommonSuffix :: (Ord k, MonoidNull v, RightGCDMonoid v) => MonoidMap k v -> MonoidMap k v -> (MonoidMap k v, MonoidMap k v, MonoidMap k v)

Strips the greatest common suffix from a pair of maps.

Given two maps m1 and m2, stripCommonSuffix produces a tuple (r1, r2, s), where:

s is the greatest common suffix of m1 and m2

r1 is the remainder of stripping suffix s from m1

r2 is the remainder of stripping suffix s from m2

The resulting suffix s can be appended to the right-hand side of either remainder r1 or r2 to reproduce either of the original maps m1 or m2 respectively:

stripCommonSuffix m1 m2
   & \(r1, _, s) -> r1 <> s == m1
stripCommonSuffix m1 m2
   & \(_, r2, s) -> r2 <> s == m2

Suffix s is identical to the result of applying commonSuffix to m1 and m2:

stripCommonSuffix m1 m2
   & \(_, _, s) -> s == commonSuffix m1 m2

Remainders r1 and r2 are identical to the results of applying stripSuffix to s and m1 or to s and m2 respectively:

stripCommonSuffix m1 m2
   & \(r1, _, s) -> Just r1 == stripSuffix s m1
stripCommonSuffix m1 m2
   & \(_, r2, s) -> Just r2 == stripSuffix s m2

This function provides the definition of stripCommonSuffix for the MonoidMap instance of RightGCDMonoid.

Examples

Expand

With String values:

>>> m1 = fromList [(1, "+++"), (2, "++a"), (3, "+aa"), (4, "aaa")]
>>> m2 = fromList [(1, "---"), (2, "--a"), (3, "-aa"), (4, "aaa")]

>>> r1 = fromList [(1, "+++"), (2, "++" ), (3, "+"  ), (4, ""   )]
>>> r2 = fromList [(1, "---"), (2, "--" ), (3, "-"  ), (4, ""   )]
>>> s  = fromList [(1,    ""), (2,   "a"), (3,  "aa"), (4, "aaa")]

>>> stripCommonSuffix m1 m2 == (r1, r2, s)
True

With Sum Natural values:

>>> m1 = fromList [("a", 0), ("b", 1), ("c", 2), ("d", 3), ("e", 4)]
>>> m2 = fromList [("a", 4), ("b", 3), ("c", 2), ("d", 1), ("e", 0)]

>>> r1 = fromList [("a", 0), ("b", 0), ("c", 0), ("d", 2), ("e", 4)]
>>> r2 = fromList [("a", 4), ("b", 2), ("c", 0), ("d", 0), ("e", 0)]
>>> s  = fromList [("a", 0), ("b", 1), ("c", 2), ("d", 1), ("e", 0)]

>>> stripCommonSuffix m1 m2 == (r1, r2, s)
True

Overlap

overlap :: (Ord k, MonoidNull v, OverlappingGCDMonoid v) => MonoidMap k v -> MonoidMap k v -> MonoidMap k v

Finds the greatest overlap of two maps.

The greatest overlap o of maps m1 and m2 is the unique greatest map that is both a suffix of m1 and a prefix of m2:

m1 == r1 <> o   
m2 ==       o <> r2

Where:

r1 is the remainder obtained by stripping suffix overlap o from m1.

(see stripSuffixOverlap)

r2 is the remainder obtained by stripping prefix overlap o from m2.

(see stripPrefixOverlap)

This function satisfies the following property:

get k (overlap m1 m2) == overlap (get k m1) (get k m2)

This function provides the definition of overlap for the MonoidMap instance of OverlappingGCDMonoid.

Examples

Expand

With String values:

>>> m1 = fromList [(1,"abc"   ), (2,"abcd"  ), (3,"abcde "), (4,"abcdef")]
>>> m2 = fromList [(1,   "def"), (2,  "cdef"), (3," bcdef"), (4,"abcdef")]
>>> m3 = fromList [(1,   ""   ), (2,  "cd"  ), (3," bcde" ), (4,"abcdef")]

>>> overlap m1 m2 == m3
True

With Sum Natural values:

>>> m1 = fromList [("a", 0), ("b", 1), ("c", 2), ("d", 3), ("e", 4)]
>>> m2 = fromList [("a", 4), ("b", 3), ("c", 2), ("d", 1), ("e", 0)]
>>> m3 = fromList [("a", 0), ("b", 1), ("c", 2), ("d", 1), ("e", 0)]

>>> overlap m1 m2 == m3
True

stripPrefixOverlap :: (Ord k, MonoidNull v, OverlappingGCDMonoid v) => MonoidMap k v -> MonoidMap k v -> MonoidMap k v

Strips from the second map its greatest prefix overlap with suffixes of the first map.

Evaluating stripPrefixOverlap m1 m2 produces the remainder r2:

m1 == r1 <> o   
m2 ==       o <> r2

Where o is the greatest overlap of maps m1 and m2: the unique greatest map that is both a suffix of m1 and a prefix of m2.

This function satisfies the following property:

get k (stripPrefixOverlap m1 m2)
    == stripPrefixOverlap (get k m1) (get k m2)

This function provides the definition of stripPrefixOverlap for the MonoidMap instance of OverlappingGCDMonoid.

Examples

Expand

With String values:

>>> m1 = fromList [(1,"abc"   ), (2,"abcd"  ), (3,"abcde" ), (4,"abcdef")]
>>> m2 = fromList [(1,   "def"), (2,  "cdef"), (3, "bcdef"), (4,"abcdef")]
>>> m3 = fromList [(1,   "def"), (2,    "ef"), (3,     "f"), (4,      "")]

>>> stripPrefixOverlap m1 m2 == m3
True

With Sum Natural values:

>>> m1 = fromList [("a", 0), ("b", 1), ("c", 2), ("d", 3), ("e", 4)]
>>> m2 = fromList [("a", 4), ("b", 3), ("c", 2), ("d", 1), ("e", 0)]
>>> m3 = fromList [("a", 4), ("b", 2), ("c", 0), ("d", 0), ("e", 0)]

>>> stripPrefixOverlap m1 m2 == m3
True

stripSuffixOverlap :: (Ord k, MonoidNull v, OverlappingGCDMonoid v) => MonoidMap k v -> MonoidMap k v -> MonoidMap k v

Strips from the second map its greatest suffix overlap with prefixes of the first map.

Evaluating stripSuffixOverlap m2 m1 produces the remainder r1:

m1 == r1 <> o   
m2 ==       o <> r2

Where o is the greatest overlap of maps m1 and m2: the unique greatest map that is both a suffix of m1 and a prefix of m2.

This function satisfies the following property:

get k (stripSuffixOverlap m2 m1)
    == stripSuffixOverlap (get k m2) (get k m1)

This function provides the definition of stripSuffixOverlap for the MonoidMap instance of OverlappingGCDMonoid.

Examples

Expand

With String values:

>>> m1 = fromList [(1,"abc"   ), (2,"abcd"  ), (3,"abcde" ), (4,"abcdef")]
>>> m2 = fromList [(1,   "def"), (2,  "cdef"), (3, "bcdef"), (4,"abcdef")]
>>> m3 = fromList [(1,"abc"   ), (2,"ab"    ), (3,"a"     ), (4,""      )]

>>> stripSuffixOverlap m2 m1 == m3
True

With Sum Natural values:

>>> m1 = fromList [("a", 0), ("b", 1), ("c", 2), ("d", 3), ("e", 4)]
>>> m2 = fromList [("a", 4), ("b", 3), ("c", 2), ("d", 1), ("e", 0)]
>>> m3 = fromList [("a", 0), ("b", 0), ("c", 0), ("d", 2), ("e", 4)]

>>> stripSuffixOverlap m2 m1 == m3
True

stripOverlap :: (Ord k, MonoidNull v, OverlappingGCDMonoid v) => MonoidMap k v -> MonoidMap k v -> (MonoidMap k v, MonoidMap k v, MonoidMap k v)

Finds the greatest overlap of two maps and strips it from both maps.

Evaluating stripOverlap m1 m2 produces the tuple (r1, o, r2), where:

m1 == r1 <> o   
m2 ==       o <> r2

Where:

o is the greatest overlap of maps m1 and m2: the unique greatest map that is both a suffix of m1 and a prefix of m2.

(see overlap)

r1 is the remainder obtained by stripping suffix overlap o from m1.

(see stripSuffixOverlap)

r2 is the remainder obtained by stripping prefix overlap o from m2.

(see stripPrefixOverlap)

This function satisfies the following property:

stripOverlap m1 m2 ==
   ( stripSuffixOverlap m2 m1
   , overlap m1 m2
   , stripPrefixOverlap m1 m2
   )

This function provides the definition of stripOverlap for the MonoidMap instance of OverlappingGCDMonoid.

GCD

gcd :: (Ord k, MonoidNull v, GCDMonoid v) => MonoidMap k v -> MonoidMap k v -> MonoidMap k v

Finds the greatest common divisor of two maps.

Satisfies the following property:

get k (gcd m1 m2) == gcd (get k m1) (get k m2)

This function provides the definition of gcd for the MonoidMap instance of GCDMonoid.

Examples

Expand

With Product Natural values, this function computes the greatest common divisor of each pair of matching values:

>>> m1 = fromList [("a", 0), ("b", 1), ("c", 2), ("d", 3)]
>>> m2 = fromList [("a", 0), ("b", 0), ("c", 0), ("d", 0)]
>>> m3 = fromList [("a", 0), ("b", 1), ("c", 2), ("d", 3)]

>>> gcd m1 m2 == m3
True

>>> m1 = fromList [("a", 0), ("b", 1), ("c", 2), ("d", 3)]
>>> m2 = fromList [("a", 2), ("b", 2), ("c", 2), ("d", 2)]
>>> m3 = fromList [("a", 2), ("b", 1), ("c", 2), ("d", 1)]

>>> gcd m1 m2 == m3
True

With Set Natural values, this function computes the set intersection of each pair of matching values:

f xs = fromList (fromList <$> xs)

>>> m1 = f [("a", [0,1,2]), ("b", [3,4,5]), ("c", [6,7,8])]
>>> m2 = f [("a", [0    ]), ("b", [  4  ]), ("c", [    8])]
>>> m3 = f [("a", [0    ]), ("b", [  4  ]), ("c", [    8])]

>>> gcd m1 m2 == m3
True

>>> m1 = f [("a", [0,1  ]), ("b", [3,4  ]), ("c", [6,7  ])]
>>> m2 = f [("a", [  1,2]), ("b", [  4,5]), ("c", [  7,8])]
>>> m3 = f [("a", [  1  ]), ("b", [  4  ]), ("c", [  7  ])]

>>> gcd m1 m2 == m3
True