Data.Total.MonoidMap.Unsafe

Description

Provides unsafe operations for the MonoidMap type.

Construction

unsafeFromMap :: Map k v -> MonoidMap k v

Unsafely constructs a MonoidMap from an ordinary Map.

Constructs a MonoidMap in constant time, without imposing the burden of a canonicalisation step to remove null values.

When applied to a given Map m, this function expects but does not check the following pre-condition:

all (not . null) m

See fromMap for a safe version of this function.